move existing logic to model layer
fr
